# laravelProject/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Gestión de Vehículos",
        "description": "API RESTful para gestionar marcas de vehículos, personas y su relación de propiedad con vehículos. Incluye operaciones CRUD completas y endpoints de relaciones.",
        "contact": {
            "email": "dumar.pabon@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de la API"
        }
    ],
    "paths": {
...
        "/api/vehiculos": {
            "get": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Lista todos los vehículos",
                "operationId": "b912521b0e41dec60b1fd4d98a0e06fe",
                "responses": {
                    "200": {
                        "description": "Lista de vehículos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Vehiculo"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Crea un nuevo vehículo",
                "operationId": "20a2f18df869f0d5e7d5ddce18285c29",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VehiculoCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vehículo creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehiculo"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/vehiculos/{id}": {
            "get": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Obtiene un vehículo por ID",
                "operationId": "e7cea4106c3b515e1fd8f8fc20a01096",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehículo encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehiculo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehículo no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Actualiza un vehículo",
                "operationId": "96084827a5107cad2b7ea8faa4eaf765",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/VehiculoUpdate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vehículo actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehiculo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehículo no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Elimina un vehículo",
                "operationId": "9b11c434f19fa00c1d0b5d3fc9ca526e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Vehículo eliminado exitosamente"
                    },
                    "404": {
                        "description": "Vehículo no encontrado"
                    }
                }
            }
        },
        "/api/vehiculos/{id}/propietarios": {
            "post": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Asigna un propietario a un vehículo",
                "operationId": "28a0b7ea8bae3b79e0482be20c878d70",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "persona_id"
                                ],
                                "properties": {
                                    "persona_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Propietario asignado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Vehiculo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Vehículo o persona no encontrada"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Vehiculo": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "modelo": {
                        "type": "string",
                        "example": "Corolla"
                    },
                    "marca_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "numero_puertas": {
                        "type": "integer",
                        "example": 4
                    },
                    "color": {
                        "type": "string",
                        "example": "Rojo"
                    },
                    "propietarios_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            },
            "VehiculoCreate": {
                "required": [
                    "modelo",
                    "marca_id",
                    "numero_puertas",
                    "color"
                ],
                "properties": {
                    "modelo": {
                        "type": "string",
                        "example": "Corolla"
                    },
                    "marca_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "numero_puertas": {
                        "type": "integer",
                        "example": 4
                    },
                    "color": {
                        "type": "string",
                        "example": "Rojo"
                    },
                    "propietarios_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            },
            "VehiculoUpdate": {
                "properties": {
                    "modelo": {
                        "type": "string",
                        "example": "Corolla"
                    },
                    "marca_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "numero_puertas": {
                        "type": "integer",
                        "example": 4
                    },
                    "color": {
                        "type": "string",
                        "example": "Rojo"
                    },
                    "propietarios_ids": {
                        "type": "array",
                        "items": {
                            "type": "integer"
                        },
                        "example": [
                            1,
                            2
                        ]
                    }
                },
                "type": "object"
            }
        }
    },
...
}